+++ src/SKIT.FlurlHttpClient.ByteDance.TikTokShop/Models/IOP/Seller/IOPSellerDistributeResponse.cs
@@ -18,7 +18,7 @@
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("supplier_shop_id")]
                         [System.Text.Json.Serialization.JsonPropertyName("supplier_shop_id")]
-                        public int SupplierShopId { get; set; }
+                        public long SupplierShopId { get; set; }
 
                         /// <summary>
                         /// 获取或设置代打订单 ID。
